Add Bootstrap template and fix inconsistancy in current template (#149)
While I was at it:
- Clean up unused imports
- Split templatetags tests from template tests
- fix example project for django 1.7 and 1.8
