+++ example/app/tables.py
@@ -1,6 +1,7 @@
 # coding: utf-8
 import django_tables2 as tables
-from .models import Country
+
+from .models import Country, Person
 
 
 class CountryTable(tables.Table):
@@ -17,3 +18,9 @@ class CountryTable(tables.Table):
 class ThemedCountryTable(CountryTable):
     class Meta:
         attrs = {'class': 'paleblue'}
+
+
+class BootstrapTable(tables.Table):
+    class Meta:
+        model = Person
+        template = 'django_tables2/bootstrap.html'
