+++ example/templates/bootstrap_template.html
@@ -1,0 +1,22 @@
+{% load render_table from django_tables2 %}
+<!doctype html>
+<html>
+<head>
+    <title>django_tables2 with bootstrap template example</title>
+
+    <link href="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.6/css/bootstrap.min.css" rel="stylesheet" />
+    <link href="{{ STATIC_URL }}django_tables2/bootstrap.css" rel="stylesheet" />
+</head>
+<body>
+    <div class="container">
+
+        <h3>django_tables2 with bootstrap template example</h3>
+
+        <div class="row">
+            <div class="col-sm-10">
+                {% render_table table %}
+            </div>
+        </div>
+    </div>
+</body>
+</html>
