+++ example/urls.py
@@ -4,7 +4,7 @@ from django.conf.urls import include, url
 from django.contrib import admin
 from django.views import static
 
-from app.views import class_based, multiple, tutorial
+from app.views import bootstrap, class_based, multiple, tutorial
 
 admin.autodiscover()
 
@@ -12,6 +12,7 @@ urlpatterns = [
     url(r'^$', multiple),
     url(r'^class-based/$', class_based),
     url(r'^tutorial/$', tutorial),
+    url(r'^bootstrap/$', bootstrap),
 
     url(r'^admin/doc/', include('django.contrib.admindocs.urls')),
     url(r'^admin/', include(admin.site.urls)),
