+++ tests/app/settings.py
@@ -1,5 +1,3 @@
-import os
-
 import six
 from django import VERSION
 from django.conf import global_settings
